wed night

+++ libft/ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: melsahha <melsahha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 17:29:42 by melsahha          #+#    #+#             */
-/*   Updated: 2022/11/11 17:37:22 by melsahha         ###   ########.fr       */
+/*   Updated: 2022/11/14 18:37:25 by melsahha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <unistd.h>
+#include "libft.h"
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	char	*p;
-	int		i;
+	size_t	i;
 
 	p = (char *) s;
 	i = 0;
